+++ src/runtime/knossos-prelude.h
@@ -6,44 +6,44 @@ namespace ks {
 // We should probably come up with a better story for the
 // tests but at the time of writing I didn't want to hold back
 // edef support any longer.
-double edef_example$af(allocator *, double x) { return x; }
-double fwd$edef_example$af(allocator *, double x, double dx) { return dx; }
-double rev$edef_example$af(allocator *, double x, double ddr) { return ddr; }
-ks::tuple<double,ks::tuple<>> suffwdpass$edef_example$af(allocator *, double x) { return ks::make_tuple(x, ks::make_tuple()); }
-double sufrevpass$edef_example$af(allocator *, double ddr, ks::tuple<>) { return ddr; }
+Float edef_example$af(allocator *, Float x) { return x; }
+Float fwd$edef_example$af(allocator *, Float x, Float dx) { return dx; }
+Float rev$edef_example$af(allocator *, Float x, Float ddr) { return ddr; }
+ks::tuple<Float,ks::tuple<>> suffwdpass$edef_example$af(allocator *, Float x) { return ks::make_tuple(x, ks::make_tuple()); }
+Float sufrevpass$edef_example$af(allocator *, Float ddr, ks::tuple<>) { return ddr; }
 
-double dot$aT1fT1f(allocator *, vec<double> const& a, vec<double> const& b)
+Float dot$aT1fT1f(allocator *, vec<Float> const& a, vec<Float> const& b)
 {
 	return ts_dot(a,b);
 }
 
-vec<double>
-mul$aT2fT1f(allocator * alloc, tensor<2, double> const& M, vec<double> const& v)
+vec<Float>
+mul$aT2fT1f(allocator * alloc, tensor<2, Float> const& M, vec<Float> const& v)
 {
 	int r = M.outer_dimension();
-	vec<double> ret(alloc, r);
+	vec<Float> ret(alloc, r);
 	for(int i = 0; i < r; ++i)
 		ret[i] = ts_dot(M[i], v);
 	return ret;
 }
 
-tuple<tensor<2, double>,vec<double>>
-rev$mul$aT2fT1f(allocator * alloc, tuple<tensor<2, double>, vec<double>> const& M_v, vec<double> const& dr)
+tuple<tensor<2, Float>,vec<Float>>
+rev$mul$aT2fT1f(allocator * alloc, tuple<tensor<2, Float>, vec<Float>> const& M_v, vec<Float> const& dr)
 {
 	auto [M, v] = M_v;
 	int r = M.outer_dimension();
 	int c = size(v);
-	tensor<2, double> retM(alloc, size(M));
+	tensor<2, Float> retM(alloc, size(M));
 	for(int i = 0; i < r; ++i) {
 		// Inlined retM[i].assign(ts_scale(dr[i], v))
-		vec<double> retrow = retM[i];
+		vec<Float> retrow = retM[i];
 		for (int j = 0; j < c; ++j)
 			retrow[j] = dr[i] * v[j];
 	}
 
-	vec<double> retv(alloc, c);
+	vec<Float> retv(alloc, c);
 	for(int i = 0; i < c; ++i) {
-		double retvi = 0;
+		Float retvi = 0;
 		for(int j = 0; j < r; ++j)
 			retvi += M.index(ks::make_tuple(j, i)) * dr[j];
 		retv[i] = retvi;
@@ -52,11 +52,11 @@ rev$mul$aT2fT1f(allocator * alloc, tuple<tensor<2, double>, vec<double>> const& 
 	return ks::make_tuple(retM,retv);
 }
 
-size_t imax$aT1f(allocator *, vec<double> const &v)
+size_t imax$aT1f(allocator *, vec<Float> const &v)
 {
     KS_ASSERT(size(v) > 0);
     size_t imax = 0;
-    double vmax = v[imax];
+    Float vmax = v[imax];
     for (int i = 1; i < size(v); ++i)
         if (v[i] > vmax)
         {
@@ -66,17 +66,17 @@ size_t imax$aT1f(allocator *, vec<double> const &v)
     return imax;
 }
 
-double max$aT1f(allocator * alloc, vec<double> const& v)
+Float max$aT1f(allocator * alloc, vec<Float> const& v)
 {
     return v[imax$aT1f(alloc, v)];
 }
 
-double rev$lgamma$af(allocator *, double x, double dr)
+Float rev$lgamma$af(allocator *, Float x, Float dr)
 {
 	std::cerr << "rev$lgamma unimp!\n" << std::endl;
 	throw "rev$gamma unimp!\n";
 }
-double fwd$lgamma$af(allocator *, double x, double dx)
+Float fwd$lgamma$af(allocator *, Float x, Float dx)
 {
   if (dx == 0.0) {
     return 0.0;
@@ -86,23 +86,23 @@ double fwd$lgamma$af(allocator *, double x, double dx)
   }
 }
 
-double pow$afi(allocator *, double x, int e)
+Float pow$afi(allocator *, Float x, int e)
 {
     return std::pow(x,e);
 }
 
-tuple<> fwd$gt(allocator *, double a,double b,double d$a,double d$b)
+tuple<> fwd$gt(allocator *, Float a,Float b,Float d$a,Float d$b)
 {
     return tuple<>();
 }
 
-tuple<double,double> rev$gt(allocator *, double a,double b, tuple<> d$r)
+tuple<Float,Float> rev$gt(allocator *, Float a,Float b, tuple<> d$r)
 {
 	std::cerr << "rev$gt unimp!\n" << std::endl;
 	throw "rev$gt unimp!\n";
 }
 
-inline double sub$aff(allocator *, double t1, double t2)
+inline Float sub$aff(allocator *, Float t1, Float t2)
 {
 	return t1 - t2;
 }
@@ -112,7 +112,7 @@ inline int sub$aii(allocator *, int t1, int t2)
 	return t1 - t2;
 }
 
-inline double div$aff(allocator *, double t1, double t2)
+inline Float div$aff(allocator *, Float t1, Float t2)
 {
 	return t1 / t2;
 }
@@ -122,7 +122,7 @@ inline int div$aii(allocator *, int t1, int t2)
 	return t1 / t2;
 }
 
-inline double neg$af(allocator *, double t)
+inline Float neg$af(allocator *, Float t)
 {
 	return -t;
 }
@@ -132,17 +132,17 @@ inline int neg$ai(allocator *, int t)
 	return -t;
 }
 
-inline double exp$af(allocator *, double d) { return exp(d); }
-inline double log$af(allocator *, double d) { return log(d); }
-inline double sin$af(allocator *, double d) { return sin(d); }
-inline double cos$af(allocator *, double d) { return cos(d); }
-inline double cosh$af(allocator *, double d) { return cosh(d); }
-inline double tanh$af(allocator *, double d) { return tanh(d); }
-inline double lgamma$af(allocator *, double d) { return lgamma(d); }
-inline double erf$af(allocator *, double d) { return erf(d); }
-inline double sqrt$af(allocator *, double d) { return sqrt(d); }
+inline Float exp$af(allocator *, Float d) { return exp(d); }
+inline Float log$af(allocator *, Float d) { return log(d); }
+inline Float sin$af(allocator *, Float d) { return sin(d); }
+inline Float cos$af(allocator *, Float d) { return cos(d); }
+inline Float cosh$af(allocator *, Float d) { return cosh(d); }
+inline Float tanh$af(allocator *, Float d) { return tanh(d); }
+inline Float lgamma$af(allocator *, Float d) { return lgamma(d); }
+inline Float erf$af(allocator *, Float d) { return erf(d); }
+inline Float sqrt$af(allocator *, Float d) { return sqrt(d); }
 
-inline double to_float$ai(allocator *, int d) { return d; }
+inline Float to_float$ai(allocator *, int d) { return d; }
 
 inline bool or$abb(allocator *, int b1, int b2)  { return b1 || b2; }
 inline bool and$abb(allocator *, int b1, int b2) { return b1 && b2; }
